refactor: extract WelcomeViewModel and split WelcomeWindowView into focused files
